Lab2

+++ lab2/favorite.js
@@ -14,13 +14,21 @@ document.addEventListener("DOMContentLoaded", () => {
         <img src="${book.img}" alt="${book.title}" class="w-full h-72 object-cover mb-3">
         <h3 class="font-semibold text-lg">${book.title}</h3>
         <p class="text-gray-600 mb-2">${book.author}</p>
-        <button class="remove-fav bg-red-500 text-white px-3 py-1 rounded hover:bg-red-600 mt-2">Remove</button>
+        <button class="remove-fav border-2 border-purple-700 p-2 m-2 rounded-lg hover:bg-purple-400 hover:border-purple-600 shadow-lg">Remove</button>
       `;
       div.querySelector(".remove-fav").addEventListener("click", () => {
-        favorites = favorites.filter(f => f.id !== book.id);
-        localStorage.setItem("favorites", JSON.stringify(favorites));
-        renderFavorites();
-      });
+        
+  const confirmRemove = confirm(`Are you sure you want to remove "${book.title}" from favorites?`);
+  if (confirmRemove) {
+    favorites = favorites.filter(f => f.id !== book.id);
+    localStorage.setItem("favorites", JSON.stringify(favorites));
+    renderFavorites();
+    alert(`"${book.title}" has been removed.`);
+  } else {
+    alert("Action canceled.");
+  }
+});
+
       favGrid.appendChild(div);
     });
   };
